Add test for ProgressCounter and MovieRepository, a couple of bugfixes

+++ src/main/java/ru/loaltyplant/movierater/model/Movie.java
@@ -11,5 +11,5 @@ import java.util.Set;
 public class Movie implements Serializable {
     private final long id;
     private final Set<Long> genreIds;
-    private final float averageRating;
+    private final double averageRating;
 }
